requirements.txt
Django==4.1.3
celery==5.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.3
uwsgi==2.0.20
django-timezone-field==5.0
drf-yasg==1.21.4
django-phonenumber-field[phonenumbers]
redis==4.4.0